LayoutTests/webanimations/frame-rate/document-timeline-maximum-frame-rate-expected.txt


PASS document.timeline.maximumFrameRate is defined and reports an expected frame rate
